requirements.txt
aniso8601==1.3.0
click==6.7
Flask==0.12.2
Flask-PyMongo==0.5.1
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pymongo==3.5.1
python-dateutil==2.6.1
python-dotenv==0.6.5
pytz==2017.2
six==1.10.0
Werkzeug==0.12.2
